change fragments name
